Modify new layout to index.html
Added a new card layout to the index.html page and changed the color theme to look better. In addition, the card layout was centralized.

Appropriate margin, border and padding were added for each content.

A new folder was created for the images and the new image was placed below the links.